+++ df-settings-clarity/src/df-settings-clarity.mjs
@@ -1,5 +1,5 @@
 /// <reference path="../../common/libWrapper.d.ts" />
-// import FuzzySearch from "./FuzzySearch.mjs";
+import { parseHTML } from "../common/fvtt.mjs";
 class DFSettingsClarity {
 	static types = ["client", "world"];
 	static patchGameSettings() {
@@ -14,18 +14,21 @@ class DFSettingsClarity {
 			pair[1].name = DFSettingsClarity.formatName(pair[1].name ?? '', pair[1]);
 		}
 	}
-	static formatName(/**@type {string}*/name, data) {
+	/**
+	 * @param {string} name
+	 * @param {object} data
+	 * @returns {string}
+	 */
+	static formatName(name, data) {
 		if (name.startsWith('👤') || name.startsWith('🌎'))
 			return name;
 		let scope;
 		if (data.scope)
-			scope = DFSettingsClarity.types.includes(data.scope)
-				? data.scope
-				: "client";
+			scope = DFSettingsClarity.types.includes(data.scope) ? data.scope : "client";
 		else if (data.restricted)
 			scope = data ? 'world' : 'client';
 		else {
-			console.warn('Unknown restriction/scope on registered setting for ' + name + '". Defaulting to "client"');
+			console.debug('Unknown restriction/scope on registered setting for ' + name + '". Defaulting to "client"');
 			scope = 'client';
 		}
 		if (scope === 'client')
@@ -34,52 +37,59 @@ class DFSettingsClarity {
 			return !name ? '🌎' : '🌎 ' + game.i18n.localize(name);
 		return name;
 	}
-	static settingsRegister(/**@type {Function}*/wrapper, /**@type {string}*/module, /**@type {string}*/key, data) {
+	/**
+	 * @param {Function} wrapper 
+	 * @param {string} module 
+	 * @param {string} key 
+	 * @param {object} data 
+	 */
+	static settingsRegister(wrapper, module, key, data) {
 		data.name = DFSettingsClarity.formatName(data.name ?? '', data);
 		wrapper(module, key, data);
 	}
-	static settingsRegisterMenu(/**@type {Function}*/wrapper, /**@type {string}*/module, /**@type {string}*/key, data) {
+	/**
+	 * @param {Function} wrapper 
+	 * @param {string} module 
+	 * @param {string} key 
+	 * @param {object} data 
+	 */
+	static settingsRegisterMenu(wrapper, module, key, data) {
 		data.name = DFSettingsClarity.formatName(data.name ?? '', data);
 		wrapper(module, key, data);
 	}
 	/** @param {Event} event */
 	static showWorldHover(event) {
-		if (event.clientX > $(event.target).offset().left + 30)
+		if (event.clientX > event.target.getBoundingClientRect().left + 30)
 			DFSettingsClarity.hideHover();
 		else
-			DFSettingsClarity.showHover($(event.target), 'world');
+			DFSettingsClarity.showHover(event.target, 'world');
 	}
 	/** @param {Event} event */
 	static showClientHover(event) {
-		if (event.clientX > $(event.target).offset().left + 30)
+		if (event.clientX > event.target.getBoundingClientRect().left + 30)
 			DFSettingsClarity.hideHover();
 		else
-			DFSettingsClarity.showHover($(event.target), 'client');
+			DFSettingsClarity.showHover(event.target, 'client');
 	}
-	static hover = $(`<div class="df-settings-clarity-tooltip" style="visibility:hidden"><span class="msg"></span><img src="data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMjAiIGhlaWdodD0iMzAiIHZlcnNpb249IjEuMSIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj48cGF0aCBkPSJtMCAwLjVhMTEuNDA4IDExLjQwOCAyNC4wNTcgMCAxIDcuNjUxNCAzLjQxNThsMTEuODQ5IDExLjA4NC0xMS44NDkgMTEuMDg0YTExLjQwOCAxMS40MDggMTU1Ljk0IDAgMS03LjY1MTQgMy40MTU4IiBmaWxsPSIjMDAwMDAwQmYiIHN0cm9rZT0iI2Y2MCIgc3Ryb2tlLWxpbmVqb2luPSJyb3VuZCIvPjwvc3ZnPg=="></div>`);
+	static hover = parseHTML(`<div class="df-settings-clarity-tooltip" style="visibility:hidden"><span class="msg"></span><img src="data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMjAiIGhlaWdodD0iMzAiIHZlcnNpb249IjEuMSIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj48cGF0aCBkPSJtMCAwLjVhMTEuNDA4IDExLjQwOCAyNC4wNTcgMCAxIDcuNjUxNCAzLjQxNThsMTEuODQ5IDExLjA4NC0xMS44NDkgMTEuMDg0YTExLjQwOCAxMS40MDggMTU1Ljk0IDAgMS03LjY1MTQgMy40MTU4IiBmaWxsPSIjMDAwMDAwQmYiIHN0cm9rZT0iI2Y2MCIgc3Ryb2tlLWxpbmVqb2luPSJyb3VuZCIvPjwvc3ZnPg=="></div>`);
+	/**
+	 * @param {HTMLElement} element
+	 * @param {string} scope
+	 */
 	static showHover(element, scope) {
 		const hover = DFSettingsClarity.hover;
-		if (hover.parent().length != 0)
+		if (hover.parentElement != null)
 			return;
-		hover.find('span.msg').text(game.i18n.localize('DF_SETTINGS_CLARITY.Scope_' + scope));
-		$(document.body).append(hover);
-		const css = {
-			left: `${element.offset().left - hover.outerWidth()}px`,
-			top: `${element.offset().top - 15}px`
-		};
-		hover.css(css);
+		hover.querySelector('span.msg').textContent = game.i18n.localize('DF_SETTINGS_CLARITY.Scope_' + scope);
+		document.body.appendChild(hover);
+		const bounds = element.getBoundingClientRect();
+		hover.style.left = `${bounds.left - hover.offsetWidth}px`;
+		hover.style.top = `${bounds.top - 15}px`;
 	}
-	/** @param {Event} event */
 	static hideHover() {
-		if (DFSettingsClarity.hover.parent().length == 0)
-			return;
 		DFSettingsClarity.hover.remove();
 	}
 }
-
-// Hooks.once('init', function () {
-// 	FuzzySearch.init();
-// });
 
 Hooks.once('setup', function () {
 	const user = game.data.users.find(x => x._id === game.userId);
@@ -88,7 +98,7 @@ Hooks.once('setup', function () {
 		DFSettingsClarity.patchGameSettings();
 		DFSettingsClarity.patchGameSettingsMenus();
 	}
-	$(document.body).append(DFSettingsClarity.hover);
+	document.body.appendChild(DFSettingsClarity.hover);
 });
 
 Hooks.once('ready', function () {
@@ -99,7 +109,7 @@ Hooks.once('ready', function () {
 		return;
 	}
 	DFSettingsClarity.hover.remove();
-	DFSettingsClarity.hover.attr("style", "");
+	DFSettingsClarity.hover.removeAttribute("style");
 });
 
 Hooks.on('renderSettingsConfig', /** @param {HTMLElement} html */ function (_app, html, _data) {